Fix FESOM datareader and int overflow  (#417)
* Fix indexing in DataReaderFesom

* Enforce using only int64 in data loading

* ruff

* ruff2

* Review

* Change int64 back to int32

+++ src/weathergen/datasets/data_reader_anemoi.py
@@ -202,7 +202,7 @@ class DataReaderAnemoi(DataReaderTimestep):
 
         return rd
 
-    def select_channels(self, ds0: anemoi_datasets, ch_type: str) -> NDArray[np.int32]:
+    def select_channels(self, ds0: anemoi_datasets, ch_type: str) -> NDArray[np.int64]:
         """
         Select source or target channels
